Remove circular dependencies from build
