+++ js/bam/mods/baseModificationConstants.js
@@ -1,0 +1,17 @@
+export const codeValues = new Map([
+    ["m", "5mC"],
+    ["h", "5hmC"],
+    ["f", "5fC"],
+    ["c", "5caC"],
+    ["g", "5hmU"],
+    ["e", "5fU"],
+    ["b", "5caU"],
+    ["a", "6mA"],
+    ["o", "8xoG"],
+    ["n", "Xao"],
+    ["C", "Unknown C"],
+    ["T", "Unknown T"],
+    ["A", "Unknown A"],
+    ["G", "Unknown G"],
+    ["N", "Unknown"]
+])
