+++ js/bam/mods/baseModificationSet.js
@@ -1,6 +1,6 @@
 import {complementBase} from "../../util/sequenceUtils.js"
 import BaseModificationKey from "./baseModificationKey.js"
-import {byteToUnsignedInt} from "./baseModificationUtils.js"
+import {codeValues} from "./baseModificationConstants.js"
 
 class BaseModificationSet {
 
@@ -32,23 +32,5 @@ class BaseModificationSet {
         return codeValues.get(this.modification) || this.modification
     }
 }
-
-const codeValues = new Map()
-codeValues.set("m", "5mC")
-codeValues.set("h", "5hmC")
-codeValues.set("f", "5fC")
-codeValues.set("c", "5caC")
-codeValues.set("g", "5hmU")
-codeValues.set("e", "5fU")
-codeValues.set("b", "5caU")
-codeValues.set("a", "6mA")
-codeValues.set("o", "8xoG")
-codeValues.set("n", "Xao")
-codeValues.set("C", "Unknown C")
-codeValues.set("T", "Unknown T")
-codeValues.set("A", "Unknown A")
-codeValues.set("G", "Unknown G")
-codeValues.set("N", "Unknown")
-
 
 export default BaseModificationSet
