+++ js/bam/mods/baseModificationUtils.js
@@ -1,29 +1,10 @@
 import BaseModificationSet from "./baseModificationSet.js"
 import {reverseComplementSequence} from "../../util/sequenceUtils.js"
-
-const codeValues = new Map([
-    ["m", "5mC"],
-    ["h", "5hmC"],
-    ["f", "5fC"],
-    ["c", "5caC"],
-    ["g", "5hmU"],
-    ["e", "5fU"],
-    ["b", "5caU"],
-    ["a", "6mA"],
-    ["o", "8xoG"],
-    ["n", "Xao"],
-    ["C", "Unknown C"],
-    ["T", "Unknown T"],
-    ["A", "Unknown A"],
-    ["G", "Unknown G"],
-    ["N", "Unknown"]
-])
-
+import {codeValues} from "./baseModificationConstants.js"
 
 function modificationName(modification) {
     return  (codeValues.has(modification)) ? codeValues.get(modification) : "Uknown"
 }
-
 
 /**
  * Parse the mm tag creating a base modification set for each modification listed.
