+++ js/bam/types.js
@@ -1,0 +1,8 @@
+/**
+ * @typedef {Object} BAMTrack
+ * @property {function} getCachedAlignmentContainers
+ * @property {function} addPairedChordsForViewport
+ * @property {function} addSplitChordsForViewport
+ */
+
+export {}
