+++ js/browser.js
@@ -27,7 +27,7 @@ import IdeogramTrack from "./ideogramTrack.js"
 import RulerTrack from "./rulerTrack.js"
 import CircularViewControl from "./ui/circularViewControl.js"
 import {createCircularView, makeCircViewChromosomes} from "./jbrowse/circularViewUtils.js"
-import ROIManager from './roi/ROIManager.js'
+import {ROIManager} from "./roi/ROIManager.js"
 import TrackROISet from "./roi/trackROISet.js"
 import SampleInfo from "./sample/sampleInfo.js"
 import HicFile from "./hic/straw/hicFile.js"
@@ -35,7 +35,7 @@ import {translateSession} from "./hic/shoeboxUtils.js"
 import Hub from "./ucsc/ucscHub.js"
 import MenuUtils from "./ui/menuUtils.js"
 import Genome from "./genome/genome.js"
-import {setDefaults} from "./igv-create.js"
+import {setDefaults} from "./util/browserConfigUtils.js"
 import {trackViewportPopoverList} from './trackViewport.js'
 import TrackBase from "./trackBase.js"
 import {loadGenbank} from "./gbk/genbankParser.js"
