+++ js/cnvpytor/cnvpytorTrack.js
@@ -4,7 +4,6 @@ import {CNVpytorVCF} from "./cnvpytorVCF.js"
 import FeatureSource from '../feature/featureSource.js'
 import {createCheckbox} from "../igv-icons.js"
 import IGVGraphics from "../igv-canvas.js"
-import VariantTrack from "../variant/variantTrack.js"
 
 class CNVPytorTrack extends TrackBase {
