+++ js/igv-create.js
@@ -26,6 +26,7 @@
 import {GoogleAuth, igvxhr} from '../node_modules/igv-utils/src/index.js'
 import Browser from "./browser.js"
 import GenomeUtils from "./genome/genomeUtils.js"
+import {setDefaults} from "./util/browserConfigUtils.js"
 
 let allBrowsers = []
 
@@ -45,7 +46,7 @@ async function createBrowser(parentDiv, config) {
         await GenomeUtils.initializeGenomes(config)
     }
 
-    setDefaults(config)
+    config = setDefaults(config)
 
     if (config.queryParametersSupported) {
         extractQuery(config)
@@ -110,89 +111,6 @@ async function visibilityChange() {
         await browser.visibilityChange()
     }
 }
-
-function setDefaults(config) {
-
-    if (undefined === config.minimumBases) {
-        config.minimumBases = 40
-    }
-
-    if (undefined === config.showIdeogram) {
-        config.showIdeogram = true
-    }
-
-    if (undefined == config.showCytobandNames) {
-        config.showCytobandNames = false
-    }
-
-    if (undefined === config.showCircularView) {
-        config.showCircularView = false
-    }
-
-    if (undefined === config.showCircularViewButton) {
-        config.showCircularViewButton = false
-    }
-
-    if (undefined === config.showTrackLabelButton) {
-        config.showTrackLabelButton = true
-    }
-
-    if (undefined === config.showTrackLabels) {
-        config.showTrackLabels = true
-    }
-
-    if (undefined === config.showCursorTrackingGuideButton) {
-        config.showCursorTrackingGuideButton = true
-    }
-
-    if (undefined === config.showCursorGuide) {
-        config.showCursorGuide = config.showCursorTrackingGuide || false   // showCursorTrackingGuide is a synonym
-    }
-
-    if (undefined === config.showCenterGuideButton) {
-        config.showCenterGuideButton = true
-    }
-
-    if (undefined === config.showCenterGuide) {
-        config.showCenterGuide = false
-    }
-
-    if (undefined === config.showSampleNames) {
-        config.showSampleNames = false
-    }
-
-    if (undefined === config.showSVGButton) {
-        config.showSVGButton = true
-    }
-
-    if (config.showControls === undefined) {
-        config.showControls = true
-    }
-
-    if (config.showNavigation === undefined) {
-        config.showNavigation = true
-    }
-
-    if (config.showRuler === undefined) {
-        config.showRuler = true
-    }
-
-    if (config.flanking === undefined) {
-        config.flanking = 1000
-    }
-
-    if (config.pairsSupported === undefined) {
-        config.pairsSupported = true
-    }
-
-    if (!config.tracks) {
-        config.tracks = []
-    }
-
-    return config
-
-}
-
 
 function extractQuery(config) {
 
@@ -282,4 +200,4 @@ async function createTrack(config, browser) {
     return await Browser.prototype.createTrack.call(browser, config)
 }
 
-export {createTrack, createBrowser, removeBrowser, removeAllBrowsers, getAllBrowsers, visibilityChange, setDefaults}
+export {createTrack, createBrowser, removeBrowser, removeAllBrowsers, getAllBrowsers, visibilityChange}
