+++ js/roi/ROIManager.js
@@ -4,8 +4,9 @@ import {getElementVerticalDimension, getFilename} from "../util/igvUtils.js"
 import {inferFileFormat} from "../util/fileFormatUtils.js"
 import ROIMenu from "./ROIMenu.js"
 import ROITable from "./ROITable.js"
+import {createRegionKey, parseRegionKey, createSelector} from './roiUtils.js'
 
-class ROIManager {
+export class ROIManager {
 
     constructor(browser) {
 
@@ -368,26 +369,5 @@ class ROIManager {
 }
 
 function locusChangeHandler() {
-    this.renderAllROISets()
+    this.updateROIRegionPositions()
 }
-
-function createRegionKey(chr, start, end) {
-    return `${chr}-${start}-${end}`
-}
-
-function createSelector(regionKey) {
-    return `[data-region="${regionKey}"]`
-}
-
-function parseRegionKey(regionKey) {
-    let regionParts = regionKey.split('-')
-    let ee = parseInt(regionParts.pop())
-    let ss = parseInt(regionParts.pop())
-    let chr = regionParts.join('-')
-
-    return {chr, start: ss, end: ee, locus: `${chr}:${ss}-${ee}`, bedRecord: `${chr}\t${ss}\t${ee}`}
-}
-
-export {createRegionKey, parseRegionKey}
-
-export default ROIManager
