+++ js/sequenceTrack.js
@@ -29,75 +29,9 @@ import {reverseComplementSequence} from "./util/sequenceUtils.js"
 import {loadSequence} from "./genome/fasta.js"
 import {defaultNucleotideColors} from "./util/nucleotideColors.js"
 import {createBlatTrack} from "./blat/blatTrack.js"
+import {translationDict} from "./util/translationUtils.js"
 
 const defaultSequenceTrackOrder = Number.MIN_SAFE_INTEGER
-
-const translationDict = {
-    'TTT': 'F',
-    'TTC': 'F',
-    'TTA': 'L',
-    'TTG': 'L',
-    'CTT': 'L',
-    'CTC': 'L',
-    'CTA': 'L',
-    'CTG': 'L',
-    'ATT': 'I',
-    'ATC': 'I',
-    'ATA': 'I',
-    'ATG': 'M',
-    'GTT': 'V',
-    'GTC': 'V',
-    'GTA': 'V',
-    'GTG': 'V',
-    'TCT': 'S',
-    'TCC': 'S',
-    'TCA': 'S',
-    'TCG': 'S',
-    'CCT': 'P',
-    'CCC': 'P',
-    'CCA': 'P',
-    'CCG': 'P',
-    'ACT': 'T',
-    'ACC': 'T',
-    'ACA': 'T',
-    'ACG': 'T',
-    'GCT': 'A',
-    'GCC': 'A',
-    'GCA': 'A',
-    'GCG': 'A',
-    'TAT': 'Y',
-    'TAC': 'Y',
-    'TAA': 'STOP',
-    'TAG': 'STOP',
-    'CAT': 'H',
-    'CAC': 'H',
-    'CAA': 'Q',
-    'CAG': 'Q',
-    'AAT': 'N',
-    'AAC': 'N',
-    'AAA': 'K',
-    'AAG': 'K',
-    'GAT': 'D',
-    'GAC': 'D',
-    'GAA': 'E',
-    'GAG': 'E',
-    'TGT': 'C',
-    'TGC': 'C',
-    'TGA': 'STOP',
-    'TGG': 'W',
-    'CGT': 'R',
-    'CGC': 'R',
-    'CGA': 'R',
-    'CGG': 'R',
-    'AGT': 'S',
-    'AGC': 'S',
-    'AGA': 'R',
-    'AGG': 'R',
-    'GGT': 'G',
-    'GGC': 'G',
-    'GGA': 'G',
-    'GGG': 'G'
-}
 
 const complement = {}
 const t1 = ['A', 'G', 'C', 'T', 'Y', 'R', 'W', 'S', 'K', 'M', 'D', 'V', 'H', 'B', 'N', 'X']
