+++ js/ui/trackMenuConstants.js
@@ -1,0 +1,7 @@
+export const colorPickerTrackTypeSet = new Set(['bedtype', 'alignment', 'annotation', 'variant', 'wig', 'interact', 'shoebox'])
+
+export const vizWindowTypes = new Set(['alignment', 'annotation', 'variant', 'eqtl', 'qtl', 'snp', 'shoebox', 'wig'])
+
+export const multiTrackSelectExclusionTypes = new Set(['sequence', 'ruler', 'ideogram'])
+
+export const autoScaleGroupColorHash = {}
